+++ rdf_differ/api/entrypoints/ui/__init__.py
@@ -12,11 +12,9 @@ Module for configuring the Flask server for UI
 
 from flask import Flask
 
-from rdf_differ.config import RDF_DIFFER_SECRET_KEY_UI
+from rdf_differ import config
 
 app = Flask(__name__)
-
-
-app.config["SECRET_KEY"] = RDF_DIFFER_SECRET_KEY_UI
+app.config["SECRET_KEY"] = config.RDF_DIFFER_SECRET_KEY_UI
 
 from . import views  # noqa: E402, F401  (Flask pattern: import views after app to register routes)