+++ tests/feature/test_diffing_two_dataset_versions.py
@@ -14,7 +14,7 @@ import pytest
 import requests
 from pytest_bdd import given, parsers, scenario, then, when
 
-from rdf_differ.config import RDF_DIFFER_FUSEKI_SERVICE
+from rdf_differ import config
 from rdf_differ.core.adapters.filesystem import dir_exists
 from rdf_differ.core.adapters.sparql import SPARQLRunner
 from rdf_differ.diffing.adapters.skos_history_wrapper import SKOSHistoryRunner
@@ -43,7 +43,7 @@ def metadata(tmpdir, files):
     metadata = {
         "basedir": tmpdir.mkdir("basedir"),
         "filename": "subdivisions-skos",
-        "endpoint": RDF_DIFFER_FUSEKI_SERVICE,
+        "endpoint": config.RDF_DIFFER_FUSEKI_SERVICE,
         "dataset": "subdiv",
         "scheme_uri": "http://publications.europa.eu/resource/authority/subdivision",
         "old_version_file": files[0],