+++ tests/unit/test_tasks.py
@@ -8,8 +8,8 @@ from unittest.mock import patch
 
 import pytest
 
+from rdf_differ import config
 from rdf_differ.api.services.celery import async_create_diff, async_generate_report
-from rdf_differ.config import RDF_DIFFER_REPORTS_DB
 from rdf_differ.core.adapters.filesystem import dir_exists
 from rdf_differ.diffing.adapters.diff_adapter import FusekiDiffAdapter, FusekiException
 
@@ -22,7 +22,12 @@ def test_async_create_diff_success(mock_inject_metadata, mock_create_diff, tmpdi
     old_version_file = cleanup_location.join("old_version.rdf")
 
     return_value = async_create_diff(
-        "dataset", {}, old_version_file, new_version_file, cleanup_location, RDF_DIFFER_REPORTS_DB
+        "dataset",
+        {},
+        old_version_file,
+        new_version_file,
+        cleanup_location,
+        config.RDF_DIFFER_REPORTS_DB,
     )
 
     mock_create_diff.assert_called_once()
@@ -48,7 +53,7 @@ def test_async_create_diff_failure(mock_inject_metadata, mock_create_diff, tmpdi
             old_version_file,
             new_version_file,
             cleanup_location,
-            RDF_DIFFER_REPORTS_DB,
+            config.RDF_DIFFER_REPORTS_DB,
         )
 
     assert not dir_exists(cleanup_location)